+++ doc/CHANGELOG.md
@@ -3,7 +3,10 @@
 HEAD:
 
 - Added light-weight polymorphic effects.
+- Defined division by zero to yield zero.
+- Added check for unconditional recursion.
 - Reworked array length from `length[x]` to `x.length`.
+- Removed switch expression.
 
 Version 0.10.0: